# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
      "ApiKeyCreateRequest" : {
        "properties" : {
          "expiresAt" : {
            "example" : "2027-01-01T00:00:00Z",
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "example" : "ci-pipeline",
            "maxLength" : 128,
            "minLength" : 0,
            "type" : "string"
          },
          "scopes" : {
            "items" : {
              "example" : "ioc:read",
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        },
        "required" : [ "name", "scopes" ],
        "type" : "object"
      },
      "ApiKeyDto" : {
        "properties" : {
          "createdAt" : {
            "example" : "2026-08-27T07:00:00Z",
            "format" : "date-time",
            "type" : "string"
          },
          "expiresAt" : {
            "example" : "2027-01-01T00:00:00Z",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "example" : "9c7a1e42-5f3b-4a10-9d2c-7e8f0a1b2c3d",
            "type" : "string"
          },
          "keyPrefix" : {
            "example" : "aB3xY9kQ",
            "type" : "string"
          },
          "lastUsedAt" : {
            "example" : "2026-08-27T08:00:00Z",
            "format" : "date-time",
            "type" : "string"
          },
          "name" : {
            "example" : "ci-pipeline",
            "type" : "string"
          },
          "revokedAt" : {
            "example" : "null",
            "format" : "date-time",
            "type" : "string"
          },
          "scopes" : {
            "items" : {
              "example" : "ioc:read",
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          }
        },
        "type" : "object"
      },
...
      "AuthResponse" : {
        "properties" : {
          "accessToken" : {
            "example" : "eyJhbGciOiJIUzI1NiJ9...",
            "type" : "string"
          },
          "expiresIn" : {
            "example" : 900,
            "format" : "int64",
            "type" : "integer"
          },
          "refreshToken" : {
            "example" : "0G2f...",
            "type" : "string"
          },
          "tokenType" : {
            "example" : "Bearer",
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/SessionUserDto"
          }
        },
        "type" : "object"
      },
      "DailyCountDto" : {
        "properties" : {
          "count" : {
            "format" : "int64",
            "type" : "integer"
          },
          "date" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldIssue"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FieldIssue" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "issue" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "IssuedApiKeyDto" : {
        "properties" : {
          "apiKey" : {
            "$ref" : "#/components/schemas/ApiKeyDto"
          },
          "key" : {
            "example" : "ctip_mvp_aB3xY9kQ7fLm2pR8sT4uV6wX0yZ1cD5e",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RefreshRequest" : {
        "properties" : {
          "refreshToken" : {
            "example" : "0G2f...",
            "maxLength" : 512,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "refreshToken" ],
        "type" : "object"
      },
...
      "SessionUserDto" : {
        "properties" : {
          "displayName" : {
            "example" : "Alice Analyst",
            "type" : "string"
          },
          "permissions" : {
            "example" : [ "ioc:read", "ioc:submit" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array",
            "uniqueItems" : true
          },
          "role" : {
            "example" : "TENANT_ADMIN",
            "type" : "string"
          },
          "tenantId" : {
            "example" : "8b1a9c33-2f4e-4d55-9f6a-0c1d2e3f4a5b",
            "type" : "string"
          },
          "userId" : {
            "example" : "3f1b0c2e-9a4d-4c1a-8b77-2b0f1a9c5d10",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "StatsSummaryDto" : {
        "properties" : {
          "byType" : {
            "additionalProperties" : {
              "format" : "int64",
              "type" : "integer"
            },
            "type" : "object"
          },
          "totalActive" : {
            "format" : "int64",
            "type" : "integer"
          },
          "trend" : {
            "items" : {
              "$ref" : "#/components/schemas/DailyCountDto"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apiKeyAuth" : {
        "description" : "Machine-to-machine key: ctip_<env>_<32 base62>",
        "in" : "header",
        "name" : "X-API-Key",
        "type" : "apiKey"
      },
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Cyber Threat Intelligence Platform read API. M1 exposes anonymous read access to public TLP:CLEAR intelligence with cursor pagination; authentication, RBAC and write endpoints arrive in M2. All errors share a single ErrorResponse shape with a log-correlated traceId (see docs/spec/09-api.md §9.4).",
    "title" : "CTIP API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/v1/api-keys" : {
      "get" : {
        "description" : "Returns every key issued for the caller's tenant, including revoked ones. The key material itself is never returned. 認證:需要 Bearer JWT 或 X-API-Key,權限 apikey:create。",
        "operationId" : "listApiKeys",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ {
                  "createdAt" : "2026-08-27T07:00:00Z",
                  "expiresAt" : null,
                  "id" : "9c7a1e42-5f3b-4a10-9d2c-7e8f0a1b2c3d",
                  "keyPrefix" : "aB3xY9kQ",
                  "lastUsedAt" : null,
                  "name" : "ci-pipeline",
                  "revokedAt" : null,
                  "scopes" : [ "ioc:read" ]
                } ],
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ApiKeyDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "API keys of the caller's tenant"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ApiKeyDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Missing apikey:create permission (FORBIDDEN)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        }, {
          "apiKeyAuth" : [ ]
        } ],
        "summary" : "List the tenant's API keys",
        "tags" : [ "API Keys" ]
      },
      "post" : {
        "description" : "Issues a key for the caller's tenant. The full key is returned exactly once and only its SHA-256 hash is stored; scopes may not exceed the creator's own permissions. 認證:需要 Bearer JWT 或 X-API-Key,權限 apikey:create。",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ApiKeyCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "apiKey" : {
                    "createdAt" : "2026-08-27T07:00:00Z",
                    "expiresAt" : null,
                    "id" : "9c7a1e42-5f3b-4a10-9d2c-7e8f0a1b2c3d",
                    "keyPrefix" : "aB3xY9kQ",
                    "lastUsedAt" : null,
                    "name" : "ci-pipeline",
                    "revokedAt" : null,
                    "scopes" : [ "ioc:read" ]
                  },
                  "key" : "ctip_mvp_aB3xY9kQ7fLm2pR8sT4uV6wX0yZ1cD5e"
                },
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedApiKeyDto"
                }
              }
            },
            "description" : "Key created; full key returned once"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedApiKeyDto"
                }
              }
            },
            "description" : "Scope exceeds creator permissions (INVALID_REQUEST)"
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IssuedApiKeyDto"
                }
              }
            },
            "description" : "Missing apikey:create permission (FORBIDDEN)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        }, {
          "apiKeyAuth" : [ ]
        } ],
        "summary" : "Create an API key",
        "tags" : [ "API Keys" ]
      }
    },
...
    "/api/v1/auth/refresh" : {
      "post" : {
        "description" : "Exchanges a refresh token for a new pair. The presented token is consumed; replaying an already-used token revokes the whole token family. 認證:以請求主體中的 refresh token 認證,不需 Authorization 標頭。",
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "accessToken" : "eyJhbGciOiJIUzI1NiJ9.e30.sig",
                  "expiresIn" : 900,
                  "refreshToken" : "0G2f8xQ...",
                  "tokenType" : "Bearer",
                  "user" : {
                    "displayName" : "Alice Analyst",
                    "permissions" : [ "ioc:read", "ioc:submit" ],
                    "role" : "TENANT_ADMIN",
                    "tenantId" : "8b1a9c33-2f4e-4d55-9f6a-0c1d2e3f4a5b",
                    "userId" : "3f1b0c2e-9a4d-4c1a-8b77-2b0f1a9c5d10"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "New session issued"
          },
          "401" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            },
            "description" : "Unknown, revoked, expired or reused token (UNAUTHENTICATED)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Rotate a refresh token",
        "tags" : [ "Auth" ]
      }
    },
...
    "/api/v1/stats/summary" : {
      "get" : {
        "description" : "Total visible ACTIVE indicators, distribution by type and the last-7-day trend (daily counts by lastSeen, zero-filled). Scoped by the caller's visibility — anonymous callers see public TLP:CLEAR data only. 認證:匿名。",
        "operationId" : "summary",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : {
                  "byType" : {
                    "DOMAIN" : 40,
                    "URL" : 38
                  },
                  "totalActive" : 231,
                  "trend" : [ {
                    "count" : 3,
                    "date" : "2026-08-20"
                  }, {
                    "count" : 0,
                    "date" : "2026-08-21"
                  } ]
                },
                "schema" : {
                  "$ref" : "#/components/schemas/StatsSummaryDto"
                }
              }
            },
            "description" : "Summary statistics"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit exceeded (RATE_LIMIT_EXCEEDED)"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected error (INTERNAL_ERROR)"
          }
        },
        "security" : [ ],
        "summary" : "Public statistics summary",
        "tags" : [ "Stats" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}